// src/ts/src/generated/geometry-type.ts
// automatically generated by the FlatBuffers compiler, do not modify

/* eslint-disable @typescript-eslint/no-unused-vars, @typescript-eslint/no-explicit-any, @typescript-eslint/no-non-null-assertion */

export enum GeometryType {
  MultiPoint = 0,
  MultiLineString = 1,
  MultiSurface = 2,
  CompositeSurface = 3,
  Solid = 4,
  MultiSolid = 5,
  CompositeSolid = 6,
  GeometryInstance = 7
}